# school-communication-platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: School Communication Platform API
  description: >
    REST API for five connected dashboards (Super Admin, Admin, Teacher,
    Student, Parent). Messaging, document sharing, announcements and
    notifications. Full reference: docs/API.md and /docs/api.html.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string, nullable: true }
        username: { type: string }
        role: { type: string, enum: [super_admin, admin, teacher, student, parent] }
        status: { type: string, enum: [active, inactive, suspended] }
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      summary: List users (super admin; admins read-only)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { type: array, items: { $ref: '#/components/schemas/User' } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Create a user (super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, username, password, role]
              properties:
                fullName: { type: string }
                username: { type: string }
                password: { type: string }
                role: { type: string }
                email: { type: string }
                phone: { type: string }
      responses:
        '201': { description: Created }
        '409': { description: Conflict }
...